upload all project
